feat(updates) redux

+++ server/routes/public.routes.js
@@ -20,26 +20,26 @@ router.get('/current-orders', async (req, res) => {
 router.post('/add-order', async (req, res) => {
   try {
     if (!req.body) {
-      res.status(400).json({ success: false, error: 'No information sent.' })
+      res.status(400).json({ success: false, error: 'No information sent.' });
       return;
     }
-  
+
     if (!req.body.order_item) {
-      res.status(400).json({ success: false, error: 'No order item sent.'});
+      res.status(400).json({ success: false, error: 'No order item sent.' });
       return;
     }
-  
+
     if (!req.body.quantity) {
-      res.status(400).json({ success: false, error: 'No quantity sent.'})
+      res.status(400).json({ success: false, error: 'No quantity sent.' });
       return;
     }
-  
+
     const orderObj = new Order({
       order_item: req.body.order_item,
       quantity: req.body.quantity,
       ordered_by: req.body.ordered_by,
     });
-  
+
     const dbResponse = await orderObj.save();
     if (dbResponse && dbResponse._id) {
       res.status(200).json({ success: true, insertedId: dbResponse._id });
@@ -53,40 +53,8 @@ router.post('/add-order', async (req, res) => {
 
 router.post('/edit-order', async (req, res) => {
   // expects id
+
   try {
-    if (!req.body.id) {
-      res.status(400).json({ success: false, error: 'No id supplied'});
-      return;
-    }
-    
-    // make sure an order exists in the database with that id
-    const targetOrder = await Order.findOne({ _id: req.body.id });
-    if (!targetOrder) {
-      res.status(400).json({ success: false, error: 'No order exists with that id!' });
-      return;
-    }
-
-    const updateResponse = await Order.updateOne({
-      _id: req.body.id
-    }, {
-      ordered_by: req.body.ordered_by,
-      order_item: req.body.order_item,
-      quantity: req.body.quantity,
-    });
-
-    if (!updateResponse || !updateResponse.nModified) {
-      res.status(400).json({ success: false, error: 'Error in database while updating' });
-      return;
-    }
-    res.status(200).json({ success: true });
-  } catch(error) {
-    res.status(500).json({ success: false, error });
-  }
-});
-
-router.post('/delete-order', async (req, res) => {
-  try {
-    // expects id
     if (!req.body.id) {
       res.status(400).json({ success: false, error: 'No id supplied' });
       return;
@@ -95,17 +63,61 @@ router.post('/delete-order', async (req, res) => {
     // make sure an order exists in the database with that id
     const targetOrder = await Order.findOne({ _id: req.body.id });
     if (!targetOrder) {
-      res.status(400).json({ success: false, error: 'No order exists with that id!' });
+      res
+        .status(400)
+        .json({ success: false, error: 'No order exists with that id!' });
+      return;
+    }
+
+    const updateResponse = await Order.updateOne(
+      {
+        _id: req.body.id,
+      },
+      {
+        ordered_by: req.body.ordered_by,
+        order_item: req.body.order_item,
+        quantity: req.body.quantity,
+      }
+    );
+
+    if (!updateResponse || !updateResponse.nModified) {
+      res
+        .status(400)
+        .json({ success: false, error: 'Error in database while updating' });
+      return;
+    }
+    const orders = await Order.find();
+    res.status(200).json({ success: true, orders });
+  } catch (error) {
+    res.status(500).json({ success: false, error });
+  }
+});
+
+router.post('/delete-order', async (req, res) => {
+  try {
+    if (!req.body.id) {
+      res.status(400).json({ success: false, error: 'No id supplied' });
+      return;
+    }
+
+    // make sure an order exists in the database with that id
+    const targetOrder = await Order.findOne({ _id: req.body.id });
+    if (!targetOrder) {
+      res
+        .status(400)
+        .json({ success: false, error: 'No order exists with that id!' });
       return;
     }
 
     const deleteResponse = await Order.deleteOne({ _id: req.body.id });
     if (!deleteResponse || !deleteResponse.n) {
-      res.status(400).json({ success: false, error: 'Unable to delete from database' });
+      res
+        .status(400)
+        .json({ success: false, error: 'Unable to delete from database' });
       return;
     }
-
-    res.status(200).json({ success: true });
+    const orders = await Order.find();
+    res.status(200).json({ success: true, orders });
   } catch (error) {
     res.status(500).json({ success: false, error });
   }
@@ -116,11 +128,12 @@ router.delete('/delete-all', async (req, res) => {
     // HITTING THIS ENDPOINT DELETES ALL ORDERS
     const deleteResponse = await Order.deleteMany({});
     if (!deleteResponse) {
-      res.status(400).json({ success: false, error: 'Error deleting all orders.' });
+      res
+        .status(400)
+        .json({ success: false, error: 'Error deleting all orders.' });
       return;
     }
     res.status(200).json({ success: true, deleted: deleteResponse.n });
-
   } catch (error) {
     res.status(500).json({ success: false, error });
   }
